Factor out BCrypt symmetric algorithm mode shared handles for sharing
