+++ src/System.Security.Cryptography.Cng/src/Internal/Cryptography/ICngSymmetricAlgorithm.cs
@@ -2,10 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Security.Cryptography;
+using Internal.NativeCrypto;
 
 namespace Internal.Cryptography
 {
@@ -30,6 +28,7 @@ namespace Internal.Cryptography
 
         // Other members.
         bool IsWeakKey(byte[] key);
+        SafeAlgorithmHandle GetEphemeralModeHandle();
     }
 }
